Fix cli commands
Something must have changed either with cliff or
with a different version of python. All of the
cli commands end with: "'Namespace' object is not subscriptable"
This is caused because of how the parsed_args fields
were accessed.

This patch addresses the issue. Now it accesses the fields
the same way as aodhclient does. The unit tests for cli
now use the arg parser to actually test this.

During this I also discovered incorrect mocking in the
configuration tests. I added another mock to the test cases
which were missing it.

Change-Id: Ib5dbbdb48bdfd3214ec76acc4c68649e25f695db