Drop direct usage of simplejson
simplejson is not in requirements so is not install automatically.
The requests library uses the built-in json module instead of the 3rd-
party simplejson library in case the simplejson library is not present.
We should use the alias (requests.JsonDecodeError) instead of
the actual exception from underlying modules to adapt to that
selection logic.

Also remove duplicated code in __str__ and __repr__. The logic in
__repr__ does not contain handling for non-json response and looks
incomplete.

Change-Id: I89749eebc1e410f023169a115ffbfa9ef04cf3ad