Additional tests and a fix for rbac code
This adds additional queries to ensure better test coverage
of the rbac code. I discovered an issue with the code, where
it wasn't able to correctly modify queries with empty curly
braces present. So for example `ceilometer_cpu{}` which is a
valid query was being modified to `ceilometer_cpu{}project={<project_id>}`
which isn't correct.

Change-Id: I95567164e4d684fb5ebb4a6558752441cab3fd93