Fix rbac mocking in unit tests
Seems like the rbac object wasn't being mocked correctly. Especialy
the disable_rbac property, which is set to False by default, but
the tests behaved as if it was set to True, which wasn't the intention.

This patch sets the property to False and fixes 2 tests, which broke
because of this. The new behavior of the tests is actually the
correct behavior in those situations.

Change-Id: I58d32d2728cf45486e54a12921834a5f3764aeb6