Send keystone tokens to Aetos only
Previously a keystone token was sent with every request, no
matter what was on the receiving end. This could apparently
pose a security threat. This patch changes the observabilityclient's
behavior to send the token only when the receiving endpoint
was discovered from the keystone endpoints and the
service_type is 'prometheus' and service_name is 'aetos'.

Change-Id: I39e84c20ef3ef4e399bf191da6b9c7ad31935fd5
Signed-off-by: Jaromir Wysoglad <jwysogla@redhat.com>