Allow for specifying rbac label name
As discussed previously on another patch, we may want to rename
the rbac label from "project" to "project_id" in the future. To
get ready for that, allow for specifying the label name when
creating the PromQLRbac object. This will allow services using
it to expose a configuration option for setting that label to
the correct name. Cloudkitty already has that option in its config.

Change-Id: I52baa0aa778c1ff5b03bfdf234323e8cda4cb278
Signed-off-by: Jaromir Wysoglad <jwysogla@redhat.com>