[tox] Use the recommended coverage configuration
The OpenStack Governance repo states that stestr should be used for running tests [1]
pytest is used for coverage and in a standalone tox env, but doesn't run the coverage tests.

There is a recommended coverage target config in the governace repo, so
this has been added in place of the pytest config

The requirement for pytest has been removed from the repo.
The cover target runs the unittests using the same config as testenv
target

[1] https://github.com/openstack/governance/blob/master/reference/pti/python.rst#python-test-running

Change-Id: I284e564ab88e249b051538dc57f7cd9cec71bc2f
Signed-off-by: Emma Foley <efoley@redhat.com>